requirements.txt
# Production
Django==3.0.1
djangorestframework==3.11.0
# Development
flake8==3.7.9
## The following requirements were added by pip freeze:
asgiref==3.2.3
entrypoints==0.3
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
sqlparse==0.3.0
